refactor and move FocusState
