+++ examples/buttons.rs
@@ -27,7 +27,6 @@ fn Button(cx: Scope<ButtonProps>) -> Element {
             width: "100%",
             height: "100%",
             background_color: "{color}",
-            // prevent_default: "false",
             tabindex: "{cx.props.layer}",
             onkeydown: |e| {
                 if let KeyCode::Space = e.data.key_code{
