+++ src/focus.rs
@@ -1,3 +1,8 @@
+use std::sync::{Arc, Mutex};
+
+use dioxus::{core::ElementId, native_core::utils::PersistantElementIter};
+
+use crate::Dom;
 use std::num::NonZeroU16;
 
 use dioxus::{
@@ -99,3 +104,124 @@ impl NodeDepState for Focus {
 
 const FOCUS_EVENTS: &[&str] = &sorted_str_slice!(["keydown", "keyup", "keypress"]);
 const FOCUS_ATTRIBUTES: &[&str] = &sorted_str_slice!(["dioxus-prevent-default", "tabindex"]);
+
+#[derive(Default)]
+pub(crate) struct FocusState {
+    pub(crate) focus_iter: Arc<Mutex<PersistantElementIter>>,
+    pub(crate) last_focused_id: Option<ElementId>,
+    pub(crate) focus_level: FocusLevel,
+}
+
+impl FocusState {
+    pub fn progress(&mut self, rdom: &mut Dom, forward: bool) -> bool {
+        if let Ok(mut focus_iter) = self.focus_iter.lock() {
+            if let Some(last) = self.last_focused_id {
+                if !rdom[last].state.focus.pass_focus {
+                    return false;
+                }
+            }
+            let mut loop_marker_id = self.last_focused_id;
+            let focus_level = &mut self.focus_level;
+            let mut next_focus = None;
+
+            loop {
+                let new = if forward {
+                    focus_iter.next(&rdom)
+                } else {
+                    focus_iter.prev(&rdom)
+                };
+                let new_id = new.id();
+                let current_level = rdom[new_id].state.focus.level;
+                if let dioxus::native_core::utils::ElementProduced::Looped(_) = new {
+                    let mut closest_level = None;
+
+                    if forward {
+                        // find the closest focusable element after the current level
+                        rdom.traverse_depth_first(|n| {
+                            let current_level = n.state.focus.level;
+                            if current_level != *focus_level {
+                                if current_level > *focus_level {
+                                    if let Some(level) = &mut closest_level {
+                                        if current_level < *level {
+                                            *level = current_level;
+                                        }
+                                    } else {
+                                        closest_level = Some(current_level);
+                                    }
+                                }
+                            }
+                        });
+                    } else {
+                        // find the closest focusable element before the current level
+                        rdom.traverse_depth_first(|n| {
+                            let current_level = n.state.focus.level;
+                            if current_level != *focus_level {
+                                if current_level < *focus_level {
+                                    if let Some(level) = &mut closest_level {
+                                        if current_level > *level {
+                                            *level = current_level;
+                                        }
+                                    } else {
+                                        closest_level = Some(current_level);
+                                    }
+                                }
+                            }
+                        });
+                    }
+
+                    // extend the loop_marker_id to allow for another pass
+                    loop_marker_id = None;
+
+                    if let Some(level) = closest_level {
+                        *focus_level = level;
+                    } else {
+                        if forward {
+                            *focus_level = FocusLevel::Unfocusable;
+                        } else {
+                            *focus_level = FocusLevel::Focusable;
+                        }
+                    }
+                }
+
+                // once we have looked at all the elements exit the loop
+                if let Some(last) = loop_marker_id {
+                    if new_id == last {
+                        break;
+                    }
+                } else {
+                    loop_marker_id = Some(new_id);
+                }
+
+                let after_previous_focused = if forward {
+                    current_level >= *focus_level
+                } else {
+                    current_level <= *focus_level
+                };
+                if after_previous_focused && current_level.focusable() {
+                    if current_level == *focus_level {
+                        next_focus = Some((new_id, current_level));
+                        break;
+                    }
+                }
+            }
+
+            if let Some((id, order)) = next_focus {
+                if order.focusable() {
+                    rdom[id].state.focused = true;
+                    if let Some(old) = self.last_focused_id.replace(id) {
+                        rdom[old].state.focused = false;
+                    }
+                    // reset the position to the currently focused element
+                    while if forward {
+                        focus_iter.next(&rdom).id()
+                    } else {
+                        focus_iter.prev(&rdom).id()
+                    } != id
+                    {}
+                    return true;
+                }
+            }
+        }
+        false
+    }
+}
